update bilby and background tests
The Bilby has the model and the updated animations (run, idle, smell) in separate files, as well as a new animator that lets it have a full or empty belly on top of the other animations, and a quick script to test the animations; it was also resized so match unity's units.
The old background tests now have the right materials; new background tests are modular but not completely satisfying; one of them has a visible outline; to test and discuss some more.
Test scenes featuring the updates are sandbox_No and sandbox_No2